+++ index.ts
@@ -1,4 +1,5 @@
 import { IncomingMessage, ServerResponse } from "http";
+import expectCt from "./middlewares/expect-ct";
 import xDnsPrefetchControl from "./middlewares/x-dns-prefetch-control";
 import xDownloadOptions from "./middlewares/x-download-options";
 import depd = require("depd");
@@ -128,7 +129,7 @@ function helmet(options: Readonly<HelmetOptions> = {}) {
 
 helmet.contentSecurityPolicy = require("helmet-csp");
 helmet.dnsPrefetchControl = xDnsPrefetchControl;
-helmet.expectCt = require("expect-ct");
+helmet.expectCt = expectCt;
 helmet.frameguard = require("frameguard");
 helmet.hidePoweredBy = require("hide-powered-by");
 helmet.hsts = require("hsts");