+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- `helmet.expectCt` is no longer a separate package. This should have no effect on end users.
+
 ## 3.23.2 - 2020-06-23
 
 ### Changed